Revert "added some try..catch for the security."
This reverts commit b6952ccfa2ee0069f9bc8a9697de6c5459c48ebe.

+++ demo/src/main/java/com/example/ClotheDAO.java
@@ -8,8 +8,8 @@ import lombok.AllArgsConstructor;
 @Getter
 @AllArgsConstructor
 public class ClotheDAO {
-    private String name;
+    private String name; 
     private BigDecimal price;
-    private short quantity;
+    private byte quantity;
     private String description;
 }
